+++ CHANGELOG.md
@@ -1,9 +1,10 @@
 ## 0.3.3
-### Bugs
-* fix tooltip offset with scrollbar on content
+* Updated MIT license
 
 ## 0.3.2
 ### Bugs
+* fix tooltip offset with scrollbar on content
+* fix issue on Firefox [#36](https://github.com/Roche/ggtips/issues/36)
 * add missing height parameter
 * enable polygon points for tooltips